scrapping_test/requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.77
botocore==1.31.77
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
deepdiff==6.7.1
dotmap==1.3.30
dynamodb-json==1.3
exceptiongroup==1.1.1
flatdict==4.0.1
h11==0.14.0
idna==3.4
jmespath==1.0.1
JSONBender==0.9.3
numpy==1.24.3
ordered-set==4.1.0
outcome==1.2.0
pandas==2.0.2
pycparser==2.21
pydash==8.0.3
PySocks==1.7.1
python-box==7.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
selenium==4.10.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.3
typing-extensions==4.12.2
tzdata==2023.3
urllib3==1.26.18
wsproto==1.2.0
